backend/requirements.txt
tensorflow==2.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
fastapi==0.95.0
pydantic==1.10.7
uvicorn>=0.24.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
matplotlib>=3.7.0
websockets>=11.0.3
